requirements.txt
streamlit
streamlit-folium
folium
numpy
Pillow
requests
